+++ R/conColScheme.R
@@ -19,11 +19,11 @@
 #' \code{colorSymbol}.
 #'
 #' @keywords utilities color
-#'
+#' @template authors-BH
 #' @export
 #'
 #' @examples
-#' if (checkForPackageWithVersion("ChemoSpec", "5.1")) {
+#' if (checkForPackageWithVersion("ChemoSpec", 6.0)) {
 #'   library("ChemoSpec")
 #'   data(metMUD1)
 #'
@@ -32,7 +32,7 @@
 #'   newSpec <- conColScheme(metMUD1, new = c("pink", "violet"))
 #' }
 #'
-#' if (checkForPackageWithVersion("ChemoSpec2D", "0.3")) {
+#' if (checkForPackageWithVersion("ChemoSpec2D", 0.5)) {
 #'   library("ChemoSpec2D")
 #'   data(MUD1)
 #'