+++ R/plotScores.R
@@ -34,48 +34,51 @@
 #' confidence ellipses.  An ellipse is drawn for each group unless there
 #' are three or fewer samples in the group.
 #'
-#' @param tol A number describing the fraction of points to be labeled.
-#' \code{tol = 1.0} labels all the points; \code{tol = 0.05} labels the most
-#' extreme 5 percent.  Set to \code{"none"} to completely suppress labels.
-#'
 #' @param use.sym A logical; if TRUE, the color scheme is set to black and the
 #' points plotted with symbols.  Applies only to \code{\link[ChemoSpec]{Spectra}} objects.
 #'
-#' @param leg.loc Character; if \code{"none"} no legend will be drawn.
-#' Otherwise, any string acceptable to \code{\link{legend}}.
+#' @template param-legloc
 #'
 #' @param \dots Additional parameters to be passed to the plotting functions.
 #'
-#' @return None.  Side effect is a plot.
+#' @template param-tol
 #'
-#' @author Bryan A. Hanson, DePauw University.
+#' @template param-graphics-return
+#'
+#' @author Bryan A. Hanson (DePauw University), Tejasvi Gupta.
 #'
 #' @keywords multivariate robust hplot
 #'
 #' @export
 #'
 #' @examples
-#' if (checkForPackageWithVersion("ChemoSpec", "5.1")) {
+#' if (checkForPackageWithVersion("ChemoSpec", 6.0)) {
 #'   library("ChemoSpec")
+#'   # This example assumes the graphics output is set to ggplot2 (see ?GraphicsOptions).
+#'   library("ggplot2")
 #'   data(metMUD1)
 #'
 #'   pca <- c_pcaSpectra(metMUD1)
-#'   plotScores(metMUD1, pca,
-#'     main = "metMUD1 NMR Data",
-#'     pcs = c(1, 2), ellipse = "cls", tol = 0.05
-#'   )
+#'   p <- plotScores(metMUD1, pca, pcs = c(1, 2), ellipse = "cls", tol = 0.05)
+#'   p <- p + ggtitle("metMUD1 NMR Data")
+#'   p
 #' }
 #'
-#' if (checkForPackageWithVersion("ChemoSpec2D", "0.3")) {
+#' if (checkForPackageWithVersion("ChemoSpec2D", 0.5)) {
 #'   library("ChemoSpec2D")
+#'   library("ggplot2")
 #'   data(MUD1)
 #'
-#'   res <- miaSpectra2D(MUD1)
-#'   plotScores(MUD1, res, main = "MIA Scores", tol = 0.1, ellipse = "cls")
+#'   mia <- miaSpectra2D(MUD1)
+#'   p1 <- plotScores(MUD1, mia, tol = 0.1, ellipse = "cls")
+#'   p1 <- p1 + ggtitle("MIA Scores")
+#'   p1
 #'
 #'   set.seed(123)
-#'   res <- pfacSpectra2D(MUD1, parallel = FALSE, nfac = 2)
-#'   plotScores(MUD1, res, tol = 0.1, leg.loc = "bottomright", main = "PARAFAC Score Plot")
+#'   pfac <- pfacSpectra2D(MUD1, parallel = FALSE, nfac = 2)
+#'   p2 <- plotScores(MUD1, pfac, tol = 0.1, leg.loc = "bottomright")
+#'   p2 <- p2 + ggtitle("PARAFAC Score Plot")
+#'   p2
 #' }
 plotScores <- function(spectra, so,
                        pcs = c(1, 2), ellipse = "none", tol = "none",