+++ R/sampleDist.R
@@ -18,7 +18,7 @@
 #'
 #' @return A numeric matrix giving the distances between the samples.
 #'
-#' @author Bryan A. Hanson, DePauw University.
+#' @template authors-BH
 #'
 #' @seealso For \code{\link[ChemoSpec]{Spectra}} objects, see \code{\link[ChemoSpec]{plotSpectraDist}}
 #' which compares all spectra to a single reference spectrum.
@@ -27,7 +27,7 @@
 #' @export
 #' @examples
 #'
-#' if (checkForPackageWithVersion("ChemoSpec", "5.1")) {
+#' if (checkForPackageWithVersion("ChemoSpec", 6.0)) {
 #'   library("ChemoSpec")
 #'   library("lattice")
 #'   data(SrE.IR)
@@ -46,7 +46,7 @@
 #'   )
 #' }
 #'
-#' if (checkForPackageWithVersion("ChemoSpec2D", "0.3")) {
+#' if (checkForPackageWithVersion("ChemoSpec2D", 0.5)) {
 #'   library("ChemoSpec2D")
 #'   library("lattice")
 #'   data(MUD1)