+++ R/scorePlot.R
@@ -1,18 +1,21 @@
 #'
 #' scorePlot
 #'
-#' @keywords multivariate robust hplot
-#' @noRd
+#' @author Bryan A. Hanson (DePauw University), Tejasvi Gupta.
 #' @export
-#' @importFrom plyr dlply llply m_ply
+#' @importFrom plyr dlply llply
+#' @importFrom ggplot2 aes_string geom_path scale_color_manual
+#' @importFrom plotly add_annotations layout
+#' @importFrom magrittr %>%
+#' @noRd
 #'
 .scorePlot <- function(spectra, so,
                        pcs = c(1, 2), ellipse = "none", tol = "none",
                        use.sym = FALSE, leg.loc = "topright", ...) {
 
-  # Handle user-provided xlim and/or ylim
+  # Save call for possible user-provided xlim and/or ylim (base graphics)
 
-  args <- as.list(match.call())[-1] # a COPY of the args for use with do.call
+  args <- as.list(match.call())[-1]
 
   # Step 0. Check the inputs
 
@@ -27,23 +30,27 @@
 
   chkSpectra(spectra)
 
-  # Prep the data
+  # Step 1. Prep the data
+
+  # For base graphics, we need to compute the ellipses *and* use them to set overall plot
+  #   limits because base graphics doesn't understand how to set limits when plotting a bunch
+  #   of different lines, points etc
+  #
+  # For ggplot2 graphics, we need to compute the ellipses but don't need to figure out any
+  #   plot limits because ggplot2 understands setting the limits for a "whole" plot.
+  #   Note that ggplot2 does not have a geom that does robust ellipses, so we have to calculate
+  #   our own ellipse data.
 
   if (case == "PCA") DF <- data.frame(so$x[, pcs], group = spectra$groups)
   if (case == "MIA") DF <- data.frame(so$C[, pcs], group = spectra$groups)
   GRPS <- dlply(DF, "group", subset, select = c(1, 2))
 
-  # Step 1.  Compute overall plot limits, incl. ellipses if requested
+  # Step 1.  Compute the ellipses if requested
 
   if ((ellipse == "cls") || (ellipse == "rob") || (ellipse == "both")) {
-    # Compute ellipses and get overall plot limits.
-    # Keep in mind the ellipses may be quite flattened and hence large.
-    # At the same time, the ellipses might be quite round and
-    # the scores well outside them, if there is an outlier.
-    # Must check all cases!
-
-    # There must be at least 3 data points per level to make a classic ellipse.
-    # Possibly more to make a robust ellipse, as at least one point may be dropped.
+    # Compute ellipses.
+    # There must be at least 3 data points per level to make a classic ellipse,
+    # more to make a robust ellipse, as at least one (outlying) point may be dropped.
 
     gr <- sumGroups(spectra)
 
@@ -55,66 +62,286 @@
 
     idx <- which(gr$no. > 3) # Index for those groups that will get ellipses
     gr <- gr[idx, ]
-    ELL <- llply(GRPS[idx], .computeEllipses) # these are the ellipses we'll need later
-
-    x.scores <- range(llply(GRPS, subset, select = 1))
-    y.scores <- range(llply(GRPS, subset, select = 2))
-    x.ell <- range(llply(ELL, function(x) {
-      range(x[1])
-    }))
-    y.ell <- range(llply(ELL, function(x) {
-      range(x[2])
-    }))
-    x.ell.r <- range(llply(ELL, function(x) {
-      range(x[4])
-    }))
-    y.ell.r <- range(llply(ELL, function(x) {
-      range(x[5])
-    }))
-    # extend.limits: stackoverflow.com/a/29647893/633251
-    x.all <- range(x.scores, x.ell, x.ell.r)
-    x.all <- x.all + diff(x.all) * 0.05 * c(-1.0, 1.15) # expand slightly for labels on right of points
-    y.all <- range(y.scores, y.ell, y.ell.r)
-    y.all <- y.all + diff(x.all) * 0.05 * c(-1.0, 1.15) # leave room for annotations at top of plot
+    ELL <- llply(GRPS[idx], .computeEllipses) # These are the ellipses we'll need later
   }
 
-  if (ellipse == "none") {
-    x.scores <- range(llply(GRPS, subset, select = 1))
-    y.scores <- range(llply(GRPS, subset, select = 2))
-    # x.all <- range(x.scores)*c(1.0, 1.15) # expand slightly for labels
-    # y.all <- range(y.scores)*c(1.0, 1.15) # leave room for annotations at top of plot
-    x.all <- range(x.scores) + diff(range(x.scores)) * 0.05 * c(-1.0, 1.15) # expand slightly for labels
-    y.all <- range(y.scores) + diff(range(y.scores)) * 0.05 * c(-1.0, 1.15) # leave room for annotations at top of plot
-  }
+  # Step 2.  Branch for each graphics mode.
 
-  # Step 2.  Draw the scores.
+  go <- chkGraphicsOpt()
 
-  dPargs <- list(PCs = DF[, 1:2], spectra = spectra, case = case, use.sym = use.sym, ... = ...)
+  # Step 2-base
 
-  # Allow user to give xlim, ylim but provide good defaults as well
-  if (!"xlim" %in% names(args)) dPargs <- c(dPargs, list(xlim = x.all))
-  if (!"ylim" %in% names(args)) dPargs <- c(dPargs, list(ylim = y.all))
+  if (go == "base") {
+    if ((ellipse == "cls") || (ellipse == "rob") || (ellipse == "both")) {
 
-  do.call(.drawPoints, dPargs)
+      # Get limits all possible pieces of the data
+      #
+      # Keep in mind the ellipses may be quite flattened and hence large.
+      # At the same time, the ellipses might be quite round and
+      # the scores well outside them, if there is an outlier.
+      # Must check all cases!
 
-  # Step 3.  Draw the ellipses if requested.
+      x.scores <- range(llply(GRPS, subset, select = 1))
+      y.scores <- range(llply(GRPS, subset, select = 2))
+      x.ell <- range(llply(ELL, function(x) {
+        range(x[1])
+      }))
+      y.ell <- range(llply(ELL, function(x) {
+        range(x[2])
+      }))
+      x.ell.r <- range(llply(ELL, function(x) {
+        range(x[4])
+      }))
+      y.ell.r <- range(llply(ELL, function(x) {
+        range(x[5])
+      }))
+      # extend.limits: stackoverflow.com/a/29647893/633251
+      x.all <- range(x.scores, x.ell, x.ell.r)
+      x.all <- x.all + diff(x.all) * 0.05 * c(-1.0, 1.15) # expand slightly for labels on right of points
+      y.all <- range(y.scores, y.ell, y.ell.r)
+      y.all <- y.all + diff(x.all) * 0.05 * c(-1.0, 1.15) # leave room for annotations at top of plot
+    }
 
-  if ((ellipse == "cls") | (ellipse == "rob") | (ellipse == "both")) .drawEllipses(ELL, gr, ellipse, use.sym, ...)
+    if (ellipse == "none") {
+      x.scores <- range(llply(GRPS, subset, select = 1))
+      y.scores <- range(llply(GRPS, subset, select = 2))
+      x.all <- range(x.scores) + diff(range(x.scores)) * 0.05 * c(-1.0, 1.15) # expand slightly for labels
+      y.all <- range(y.scores) + diff(range(y.scores)) * 0.05 * c(-1.0, 1.15) # leave room for annotations at top of plot
+    }
 
-  # Step 4.  Decorations
+    # Now we have our limits, plot the scores after accounting for user provided xlim, ylim
 
-  if (case == "PCA") {
-    .addMethod(so)
-    if (leg.loc != "none") .addLegend(spectra, leg.loc, use.sym, bty = "n")
-    .addEllipseInfo(ellipse)
-  }
+    dPargs <- list(PCs = DF[, 1:2], spectra = spectra, case = case, use.sym = use.sym, ... = ...)
 
-  if (case == "MIA") {
-    if (leg.loc != "none") .addLegend(spectra, leg.loc, use.sym = FALSE, bty = "n")
-    .addEllipseInfo(ellipse)
-  }
+    # Allow user to give xlim, ylim but provide good defaults as well
+    if (!"xlim" %in% names(args)) dPargs <- c(dPargs, list(xlim = x.all))
+    if (!"ylim" %in% names(args)) dPargs <- c(dPargs, list(ylim = y.all))
 
-  # Step 5.  Label extremes if requested
+    do.call(.drawPoints, dPargs)
 
-  if (tol != "none") .labelExtremes(DF[, 1:2], spectra$names, tol)
+    # Draw the ellipses if requested.
+
+    if ((ellipse == "cls") | (ellipse == "rob") | (ellipse == "both")) .drawEllipses(ELL, gr, ellipse, use.sym, ...)
+
+    # Do the decorations
+
+    if (case == "PCA") {
+      .addMethod(so)
+      if (all(leg.loc != "none")) {
+        leg.loc <- .prepLegendCoords(go, leg.loc, x.all[1], x.all[2], y.all[1], y.all[2])
+        .addLegend(spectra, leg.loc, use.sym, bty = "n")
+      }
+      .addEllipseInfo(ellipse)
+    }
+
+    if (case == "MIA") {
+      if (all(leg.loc != "none")) {
+        leg.loc <- .prepLegendCoords(go, leg.loc, x.all[1], x.all[2], y.all[1], y.all[2])
+        .addLegend(spectra, leg.loc, use.sym = FALSE, bty = "n")
+      }
+      .addEllipseInfo(ellipse)
+    }
+
+    # Label extremes if requested
+
+    if (tol != "none") .labelExtremes(DF[, 1:2], spectra$names, tol)
+  } # end of go == "base"
+
+  # Step 2-ggplot2
+
+  if ((go == "ggplot2") || (go == "plotly")) {
+    args <- as.list(match.call()[-1]) # Capturing xlabel and ylabel from plotScore call
+    xlab <- eval(args$xlab)
+    ylab <- eval(args$ylab)
+    .chkReqGraphicsPkgs("ggplot2")
+
+    x <- y <- name <- label <- NULL # satisfy CRAN check engine
+
+    ## Prepare the main plot
+
+    if (case == "PCA") {
+      if (!use.sym) {
+        p <- ggplot(DF) +
+          geom_point(aes_string(x = colnames(DF)[1], y = colnames(DF)[2]), color = spectra$colors, shape = 20, size = 3) +
+          labs(x = xlab, y = ylab)
+      }
+
+      if (use.sym) {
+        p <- ggplot(DF) +
+          geom_point(aes_string(x = colnames(DF)[1], y = colnames(DF)[2]), color = "black", shape = spectra$sym) +
+          labs(x = xlab, y = ylab)
+      }
+
+      if (go == "ggplot2") {
+        p <- p + .ggAnnotate(so$method, x = 0.05, y = 0.98, just = "left", gp = gpar(fontsize = 8))
+      }
+    } # end of case == "PCA"
+
+    if (case == "MIA") {
+      p <- ggplot(DF) +
+        geom_point(aes_string(x = colnames(DF)[1], y = colnames(DF)[2]), color = spectra$colors, shape = 20, size = 3)
+    }
+
+    # Change theme to theme_bw() and remove grids
+    p <- p + theme_bw() +
+      theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank())
+
+    ## Take care of the ellipse options -- classical
+
+    # Bring in helper function if needed
+    if ((ellipse == "cls") | (ellipse == "rob") | (ellipse == "both")) {
+      .ggPrepEllipseCoords <- function(data) {
+        df <- data.frame(x = numeric(), y = numeric(), name = character())
+        for (i in 1:length(data)) {
+          x <- c()
+          y <- c()
+          name <- c()
+          total <- length(data[[i]]) / 2
+          for (j in 1:total) {
+            x <- c(x, data[[i]][j, 1])
+            y <- c(y, data[[i]][j, 2])
+          }
+          name <- rep(names(data)[i], total)
+          temp <- data.frame(x, y, name)
+          df <- rbind(df, temp)
+        }
+        return(df)
+      }
+    }
+
+    if (ellipse == "cls") {
+      cls.coords <- llply(ELL, function(x) {
+        x[1:2]
+      })
+      cls.coords <- llply(cls.coords, function(x) {
+        do.call(cbind, x)
+      })
+      df.cls <- .ggPrepEllipseCoords(cls.coords)
+
+      if (!use.sym) {
+        p <- p + geom_path(data = df.cls, aes(x = x, y = y, color = name), linetype = 2) +
+          scale_color_manual(values = gr$color)
+      }
+
+      if (use.sym) {
+        color.black <- rep("black", length(gr$color))
+        p <- p + geom_path(data = df.cls, aes(x = x, y = y, color = name), linetype = 2) +
+          scale_color_manual(values = color.black)
+      }
+
+      if (go == "ggplot2") {
+        p <- p + .ggAnnotate("cls")
+      }
+    }
+
+    ## Take care of the ellipse options -- robust
+
+    if (ellipse == "rob") {
+      rob.coords <- llply(ELL, function(x) {
+        x[4:5]
+      })
+      rob.coords <- llply(rob.coords, function(x) {
+        do.call(cbind, x)
+      })
+      df.rob <- .ggPrepEllipseCoords(rob.coords)
+
+
+      if (!use.sym) {
+        p <- p + geom_path(data = df.rob, aes(x = x, y = y, color = name)) +
+          scale_color_manual(values = gr$color)
+      }
+
+      if (use.sym) {
+        color.black <- rep("black", length(gr$color))
+        p <- p + geom_path(data = df.rob, aes(x = x, y = y, color = name)) +
+          scale_color_manual(values = color.black)
+      }
+
+      if (go == "ggplot2") {
+        p <- p + .ggAnnotate("rob")
+      }
+    }
+
+    ## Take care of the ellipse options -- both classical and robust
+
+    if (ellipse == "both") {
+      cls.coords <- llply(ELL, function(x) {
+        x[1:2]
+      })
+      cls.coords <- llply(cls.coords, function(x) {
+        do.call(cbind, x)
+      })
+      df.cls <- .ggPrepEllipseCoords(cls.coords) # Data frame with cls.coords values
+
+      rob.coords <- llply(ELL, function(x) {
+        x[4:5]
+      })
+      rob.coords <- llply(rob.coords, function(x) {
+        do.call(cbind, x)
+      })
+      df.rob <- .ggPrepEllipseCoords(rob.coords) # Data frame with rob.coords values
+
+
+      if (!use.sym) {
+        lines <- rep(2, length(gr$color))
+        p <- p + geom_path(data = df.cls, aes(x = x, y = y, color = name), linetype = 2)
+
+        p <- p + geom_path(data = df.rob, aes(x = x, y = y, color = name)) +
+          scale_color_manual(values = gr$color)
+      }
+
+      if (use.sym) {
+        color.black <- rep("black", length(gr$color))
+        p <- p + geom_path(data = df.cls, aes(x = x, y = y, color = name))
+
+        p <- p + geom_path(data = df.rob, aes(x = x, y = y, color = name), linetype = 2) +
+          scale_color_manual(values = color.black)
+      }
+
+
+      if (go == "ggplot2") {
+        p <- p + .ggAnnotate("both")
+      }
+    }
+
+    ## Final touches
+
+    if (go == "ggplot2") {
+      # label extremes
+      if (tol != "none") {
+        CoordList <- .getExtremeCoords(DF[, 1:2], spectra$names, tol)
+        df <- data.frame(x = CoordList$x, y = CoordList$y, label = CoordList$l)
+        p <- p + .ggRepel(df)
+      }
+
+      # removing the ggplot legend & adding our own
+      p <- p + theme(legend.position = "none")
+      if (all(leg.loc != "none")) {
+        p <- p + .ggAddLegend(spectra, use.sym, leg.loc)
+      }
+
+      return(p)
+    } else {
+      .chkReqGraphicsPkgs("plotly")
+      p <- ggplotly(p, tooltip = c(colnames(DF[1]), colnames(DF[2]), "name"))
+      if (tol != "none") {
+        CoordList <- .getExtremeCoords(DF[, 1:2], spectra$names, tol)
+        df <- data.frame(x = CoordList$x, y = CoordList$y, label = CoordList$l)
+        p <- p %>% add_annotations(
+          x = df$x, y = df$y, text = df$label, xref = "x",
+          yref = "y",
+          showarrow = TRUE,
+          arrowhead = 4,
+          arrowsize = .5,
+          ax = 40,
+          ay = -25,
+          font = list(
+            size = 12
+          )
+        )
+      }
+      p <- p %>% layout(showlegend = FALSE)
+      return(p)
+    }
+  } # end of go == "ggplot2" and go == "plotly"
 } # End of plotScores